Add-location-support (#5)
* chore: Refactor league data synchronization jobs and add factory bot

Refactor the league data synchronization jobs to improve code organization and readability. This includes renaming the `LeagueScheduleDataService` to `LeagueScheduleSyncService` and updating the corresponding references. The `PullLeagueScheduleDataJob` has been removed and replaced with the `LeagueScheduleSyncJob` which now calls the `LeagueScheduleSyncService.sync` method. The `refresh_games` and `refresh_teams` actions in the `LeaguesController` have been removed as they are no longer needed. The `SyncAllOrganizationsJob` now uses a block instead of `find_each` to perform organization leagues synchronization. Additionally, the `League` model now calls `LeagueScheduleSyncJob.perform_later` in the `sync_schedule` method.

* Refactor league data synchronization jobs and add location path to organizations

Refactor the league data synchronization jobs to improve code organization and readability. This includes renaming the `LeagueScheduleDataService` to `LeagueScheduleSyncService` and updating the corresponding references. The `PullLeagueScheduleDataJob` has been removed and replaced with the `LeagueScheduleSyncJob` which now calls the `LeagueScheduleSyncService.sync` method. The `refresh_games` and `refresh_teams` actions in the `LeaguesController` have been removed as they are no longer needed. The `SyncAllOrganizationsJob` now uses a block instead of `find_each` to perform organization leagues synchronization. Additionally, the `League` model now calls `LeagueScheduleSyncJob.perform_later` in the `sync_schedule` method. This commit also adds a new migration to add the `location_path` column to the `organizations` table.

Note: This commit message follows the established convention of using the prefix "refactor" for code restructuring and provides a clear summary of the code changes made.

* chore: Update Gemfile and Gemfile.lock to add geocoder gem

* add geocoding on location model

* add league locations sync job

* add location to league schedule service

* add location and description to event details